itmo_advisor/requirements.txt
# requirements.txt
requests
beautifulsoup4
pdfminer.six
python-docx
scikit-learn
python-telegram-bot==21.6
python-dotenv